chore: unbuild circular dependency
Inline `ResizeUtil` into behavior to prevent circular depdendency.

It is not being used outside the behavior anyway and uses behavior
internals (lane dimensions).
