+++ test/spec/features/modeling/behavior/ResizeBehavior.utility.lanes.bpmn
